% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_accounting_area.R
\docType{data}
\name{accounting_area}
\alias{accounting_area}
\title{Example accounting area polgygons for eaTools}
\format{
An sf object
}
\usage{
accounting_area
}
\description{
Short description of data . . .
}
\keyword{datasets}
