% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_polygons}
\alias{ex_polygons}
\title{ex_polygons}
\format{
An sf object
\describe{
\item{ID}{Unique row IDs}
\item{SHAPE_Area}{The area of the polygons}
\item{SHAPE}{The geometry of the sf object}
\item{condition_variable_1}{A positive number between zero and 10}
\item{condition_variable_2}{A normally distributed vector with mean =2 and sd = 2}
\item{condition_variable_3}{A number with value 0 or 1}
}
}
\usage{
ex_polygons
}
\description{
Example polygon data for eaTools. ex_polygons is a sf object with a dummy condition variable (un-scaled). It does not cover the entire accounting area.
}
\keyword{datasets}
