% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_raster}
\alias{ex_raster}
\title{ex_raster}
\format{
A stars object with 40 rows and 34 column(s)
}
\usage{
ex_raster
}
\description{
ex_raster is a 1000x1000m raster grid over the extent of the account area, with cells randomly classed as either 1, 2 or 3.
}
\keyword{datasets}
