% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accounting_area}
\alias{accounting_area}
\title{accounting_area}
\format{
An sf object
\describe{
\item{name}{Municipality name}
\item{SHAPE_Area}{The area of the polygons}
\item{geometry}{The geometry of the sf object}
}
}
\usage{
accounting_area
}
\description{
accounting_area is the outline of five regions.
}
\keyword{datasets}
