% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_homogeneous_area.R
\name{ea_homogeneous_area}
\alias{ea_homogeneous_area}
\title{ea_homogeneous_area}
\usage{
ea_homogeneous_area(delineation1 = NULL, delineation2 = NULL)
}
\arguments{
\item{delineation1}{A geometry (sf or stars object)}

\item{delineation2}{A second, optional, geometry (sf or stars object)}
}
\value{
The function returns an sf object with polygons defined by the intersection of \code{delineation1} and \code{delineation2}. If either of the inputs are rasters, they will be vectorised, and neighboring cells with identical values will be merged.
}
\description{
A function to create a new geometry based on two input geometries. The function is meant to precede \code{ea_spread} and to define homogeneous area over which we can spread (i.e. aggregate) indicator values.
}
\examples{
data("ex_raster")
ea_homogeneous_area(ex_raster)
}
