% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_normalise.R
\name{ea_normalise}
\alias{ea_normalise}
\title{ea_normalise}
\usage{
ea_normalise(x, scaling_function = "linear")
}
\arguments{
\item{x}{numerical vector to be normalised}

\item{scaling_function}{one of c("linear", "exponential)}
}
\value{
numerical vector, a normalised version of \code{x}
}
\description{
A function to normalise, or re-scale, a numerical vector such as a condition to become a value between 0 and 1, bound by an upper reference level.
}
\examples{
data("ex_polygons")
ea_normalise(ex_polygons$condition_variable_1)
}
