% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_polygons.R
\docType{data}
\name{ex_polygons}
\alias{ex_polygons}
\title{Example polgying data for eaTools}
\format{
An sf object
\describe{
\item{column1}{Describe column here . . .}
\item{column2}{Describe column here . . .}
}
\usage{
ex_polygons
}
\description{
Short description of data . . .
}
\keyword{datasets}
